Add files via upload
This is a version of the SD card sensor data recorder with accurate time from a DS3231 timer. The SD card and sensors are set up as before and the wiring for the DS3231 timer is indicated in a comment at the top of the file. Additionally, it requires two libraries for the timer functions to work: DS3231 by Andrew Wickert, Eric Ayars, Jean-Claude Wippler and RTCtime by smz. The RTCtime library is used to initialize the timer to the correct local time, make sure the time zone is set to 0 in the #define MY_TIMEZONE statement at the top of the DS3231_Alarms_And_Temperature program. The DS3231 library provides the functions to read the time which are used to record the time to the SD card. You will need to include: #include <Wire.h>, #include "DS3231.h", and RTClib RTC; at the top of the program. You will also put Wire.begin() in the setup(); Put DateTime now = RTC.now(); at the start of loop() and then calls for time will be of the form:  Serial.print(now.year(), DEC); or myFile.print(now.year(), DEC); with "year" replaced with month, day, hour, minute, and second respectively. This should just print the value for that time increment in a decimal format.